feat(channel): promote Feishu channel provider (#209 slice 5)
Promote @excitedjs/feishu-channel from scaffold to the built-in ChannelProvider package while preserving builtin:feishu behavior and core-owned routing policy.
