+++ common/changes/@excitedjs/dreamux/i209-feishu-channel-provider_2026-06-13-00-00.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Move the built-in Feishu channel implementation into the publishable @excitedjs/feishu-channel package and depend on it by default (issue #209 slice 5). channel/feishu/ becomes a thin core-owned adapter that resolves the host contracts the package must not reconstruct (bot secret/app id from the dispatcher row+config, the per-dispatcher state dir for access.json/chat-bots.json, the attachment cache dir) and constructs the package session; the Channel MCP server descriptor and admin-method routing stay in core, and bot/feishu-mcp-surface become re-export shims so existing import paths are stable. Routing, binding state, authorization, and Team lifecycle stay core-owned. The builtin:feishu alias and current operator/runtime behavior are unchanged; no operator action required on upgrade.",
+      "type": "patch",
+      "packageName": "@excitedjs/dreamux"
+    }
+  ],
+  "packageName": "@excitedjs/dreamux",
+  "email": "053700@gmail.com"
+}
