+++ common/changes/@excitedjs/feishu-channel/i209-feishu-channel-provider_2026-06-13-00-00.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "Promote @excitedjs/feishu-channel from scaffold to the publishable built-in Feishu ChannelProvider package (alias builtin:feishu, issue #209 slice 5). It now owns the live Feishu channel session, access/trust behavior, inbound normalization, attachment handling, and MCP tool backing, implementing the neutral @excitedjs/dreamux-types ChannelProvider/ChannelSession contract on top of @excitedjs/feishu-transport. Depends on @excitedjs/dreamux-types + @excitedjs/feishu-transport only — never on @excitedjs/dreamux core. The host supplies the bot secret/app id, the per-dispatcher state dir, and the attachment cache dir through the session options; the Channel MCP server descriptor and admin routing stay core-owned.",
+      "type": "minor",
+      "packageName": "@excitedjs/feishu-channel"
+    }
+  ],
+  "packageName": "@excitedjs/feishu-channel",
+  "email": "053700@gmail.com"
+}
