+++ packages/channel/feishu-channel/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@excitedjs/feishu-channel",
-  "version": "0.0.0",
-  "description": "dreamux-side Feishu channel layer (inbound filter + conversation→thread mapping + forwarding) on top of @excitedjs/feishu-transport. PR0 scaffold (issue excitedjs/dreamux#25).",
+  "version": "0.1.0",
+  "description": "Built-in Feishu ChannelProvider package for Dreamux (alias builtin:feishu). Owns Feishu channel session logic, inbound normalization, access/trust, and MCP tool backing on top of @excitedjs/feishu-transport; depends on @excitedjs/dreamux-types + @excitedjs/feishu-transport only, never on @excitedjs/dreamux core (issue excitedjs/dreamux#209).",
   "license": "MIT",
   "type": "module",
   "main": "./dist/index.js",
@@ -18,7 +18,8 @@
   },
   "files": [
     "dist",
-    "README.md"
+    "README.md",
+    "LICENSE"
   ],
   "scripts": {
     "build": "tsc -p tsconfig.json",
@@ -30,6 +31,7 @@
     "prepublishOnly": "npm run build"
   },
   "dependencies": {
+    "@excitedjs/dreamux-types": "workspace:*",
     "@excitedjs/feishu-transport": "workspace:*"
   },
   "devDependencies": {
