+++ packages/channel/feishu-channel/src/feishu-message.ts
@@ -9,7 +9,7 @@ import {
 import { basename, resolve } from 'node:path';
 import type { Readable } from 'node:stream';
 
-import { ensureOwnerOnlyDir } from '../../platform/owner-only-dir.js';
+import { ensureOwnerOnlyDir } from './internal/os.js';
 
 import type {
   FeishuMessageResourceFetcher,
