+++ packages/dreamux/CLAUDE.md
@@ -26,7 +26,7 @@ Two settled shape rules govern where code lives:
 | `dispatcher-service/` | the Dispatcher Service entity — see [`dispatcher-service/CLAUDE.md`](src/dispatcher-service/CLAUDE.md) | holds the dispatcher agent + orchestrates teammates |
 | `dispatcher-service/dispatcher/` | `DispatcherAgentService` (slots / start / resume / stop / restart-notice / channel session / role MCP injection) + dispatcher base prompt | dispatcher agent lifecycle is tied to the server |
 | `dispatcher-service/teammate/` | `TeamMateAgentService` + identity-store + runtime-state + types + teammate MCP descriptor | agent-centric teammates (no `task`): spawn/send/close + forward-only history (send reopens a closed teammate; no separate `resume` verb, #155) |
-| `channel/feishu/` | the built-in Feishu bidirectional channel: bot, session, gate, message, mcp-surface, chat-bots, introduce | Feishu owns its MCP end-to-end; it is **not** a registry provider |
+| `channel/feishu/` | the core-owned `builtin:feishu` **adapter** over the published `@excitedjs/feishu-channel` package (#209 slice 5): `feishu-channel.ts` (host→package adapter: resolves bot secret/app id + state/cache dirs, plus the sessionless `list_chat_bots` host helper), `feishu-mcp-surface.ts` (host-owned MCP server descriptor + admin-method routing; re-exports the package tool parser), `bot.ts` (re-export shim); the engine (session/bot/gate/message/chat-bots/introduce/tool-parsing) lives in the package | the package owns Feishu platform I/O + tool backing and never imports core; core keeps routing/binding/auth/the MCP shim |
 | `channel/plugin.ts` | TS interface reservation for future subscription-style channel plugins (github/jira) | interface-only this phase; not loaded or run |
 | `registry/` | provider registry/loader + provider-ref grammar | resolves `builtin:` / `npm:` refs; exactly two kinds: `channel`, `agentRuntime` |
 | `mcp/` | stdio MCP shim processes (`feishu-mcp`, `teammate-mcp`) | thin JSON-RPC bridges that forward to the admin socket |
