+++ packages/dreamux/tests/codex-live.test.ts
@@ -46,7 +46,8 @@ import {
   type FakeFeishuBot,
   type FeishuInboundEvent,
 } from '../src/channel/feishu/bot.js';
-import { saveDispatcherAccess } from '../src/channel/feishu/feishu-gate.js';
+import { saveDispatcherAccess } from '@excitedjs/feishu-channel';
+import { dispatcherDir } from '../src/platform/paths.js';
 import type { DreamuxConfig } from '../src/config/config.js';
 import type {
   ServerNotification,
@@ -447,7 +448,7 @@ describe('codex live integration', () => {
       // Onboard the live sender onto the global allow-user list so the folded
       // group messages are delivered (empty `allow_users` authorizes nobody
       // under the follow-user gate).
-      await saveDispatcherAccess('live', {
+      await saveDispatcherAccess(dispatcherDir('live'), {
         version: 2,
         allow_users: ['sender-live'],
         group: { policy: 'follow-user', allow_chats: [], require_mention: true },
